requirements.txt
Django>=4.2
djangorestframework
mysqlclient
opensearch-py>=2.1
gunicorn
django-cors-headers
Pillow>=9.0.0
Faker>=19.0
factory_boy>=3.3
opensearch-py>=2.4.0
celery>=5.3.6